Move zipping logic to separate package
